fix: cast numpy ufunc returns to satisfy mypy no-any-return
np.clip() and np.tanh() return Any under mypy's numpy stubs; chain
.astype(float) so the declared return types are satisfied.

+++ src/amazon_utac/crep_amazon.py
@@ -101,7 +101,7 @@ class AmazonCREP:
         gives E ≈ 0.25, contributing to Γ ≈ 0.116.
         """
         synergy = np.asarray(deforestation_fraction) * np.asarray(drought_anomaly)
-        return np.tanh(synergy / 0.50)  # scale: synergy=0.50 → E≈0.76; 0.12→E≈0.24
+        return np.tanh(synergy / 0.50).astype(float)  # scale: synergy=0.50 → E≈0.76; 0.12→E≈0.24
 
     def component_P(
         self,
